tests: move CompiledContractCache errors to a dedicated module
